% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countdate.R
\name{daysto}
\alias{daysto}
\title{Calculate days left to a certain date from today}
\usage{
daysto(date)
}
\arguments{
\item{date}{format:YYYY-MM-DD}
}
\value{
Difference of days between today and the date you've entered
}
\description{
Calculate days left to a certain date from today
}
\examples{
daysto("2040-01-01")
}
