% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countdate.R
\name{startdate}
\alias{startdate}
\title{The earliest and latest starting date for your OPT}
\usage{
startdate(date)
}
\arguments{
\item{date}{format:YYYY-MM-DD}
}
\value{
The earliest and latest starting date possible, based on your graduation date
}
\description{
The earliest and latest starting date for your OPT
}
\examples{
startdate("2022-05-01") #Please refer to your own program for the graduation/completion date
}
