% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countdate.R
\name{useddays}
\alias{useddays}
\title{How many days you have used in your current academic level of OPT}
\usage{
useddays(begin, end)
}
\arguments{
\item{begin}{format:YYYY-MM-DD}

\item{end}{format:YYYY-MM-DD}
}
\value{
Total amount of days that were used
}
\description{
How many days you have used in your current academic level of OPT
}
\examples{
useddays("2021-05-01","2021-08-01") #If you had used OPT at your current academic level of OPT previously
}
