% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countdate.R
\name{whentoapply}
\alias{whentoapply}
\title{The earliest and latest you shall apply for your initial OPT}
\usage{
whentoapply(date)
}
\arguments{
\item{date}{format:YYYY-MM-DD}
}
\value{
The day range for OPT application, based on your graduation date
}
\description{
The earliest and latest you shall apply for your initial OPT
}
\examples{
whentoapply("2022-05-01") #Please refer to your own program for the graduation/completion date
}
